+++ entities/conversation-memory/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @brains/conversation-memory
 
+## 0.2.0-alpha.356
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/dashboard@0.2.0-alpha.356
+  - @brains/content-formatters@0.2.0-alpha.356
+  - @brains/contracts@0.2.0-alpha.356
+  - @brains/ui-library@0.2.0-alpha.356
+  - @brains/utils@0.2.0-alpha.356
+  - @brains/conversation-service@0.2.0-alpha.356
+  - @brains/plugins@0.2.0-alpha.356
+  - @brains/templates@0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ### Patch Changes