+++ entities/document/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @brains/document-plugin
 
+## 0.2.0-alpha.356
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.356
+  - @brains/document@0.2.0-alpha.356
+  - @brains/media-renderer@0.2.0-alpha.356
+  - @brains/utils@0.2.0-alpha.356
+  - @brains/plugins@0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ### Patch Changes