+++ interfaces/chat-repl/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @brains/chat-repl
 
+## 0.2.0-alpha.356
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/utils@0.2.0-alpha.356
+  - @brains/job-queue@0.2.0-alpha.356
+  - @brains/plugins@0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ### Patch Changes