+++ plugins/admin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @brains/admin
 
+## 0.2.0-alpha.356
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/utils@0.2.0-alpha.356
+  - @brains/auth-service@0.2.0-alpha.356
+  - @brains/plugins@0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ### Patch Changes