+++ plugins/atproto/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @brains/atproto
 
+## 0.2.0-alpha.356
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/atproto-contracts@0.2.0-alpha.356
+  - @brains/contracts@0.2.0-alpha.356
+  - @brains/utils@0.2.0-alpha.356
+  - @brains/plugins@0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ### Patch Changes