+++ plugins/site-info/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @brains/site-info
 
+## 0.2.0-alpha.356
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.356
+  - @brains/site-composition@0.2.0-alpha.356
+  - @brains/utils@0.2.0-alpha.356
+  - @brains/entity-service@0.2.0-alpha.356
+  - @brains/plugins@0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ### Patch Changes