+++ plugins/studio/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @brains/studio
 
+## 0.2.0-alpha.356
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/app-ui-react@0.2.0-alpha.356
+  - @brains/console-theme@0.2.0-alpha.356
+  - @brains/contracts@0.2.0-alpha.356
+  - @brains/operator-view-react@0.2.0-alpha.356
+  - @brains/utils@0.2.0-alpha.356
+  - @brains/auth-service@0.2.0-alpha.356
+  - @brains/plugins@0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ### Patch Changes