+++ shared/console-theme/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brains/console-theme
 
+## 0.2.0-alpha.356
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/theme-base@0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ### Patch Changes