+++ shared/content-formatters/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @brains/content-formatters
 
+## 0.2.0-alpha.356
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.356
+  - @brains/utils@0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ### Patch Changes