+++ shared/site-composition/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @brains/site-composition
 
+## 0.2.0-alpha.356
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/content-formatters@0.2.0-alpha.356
+  - @brains/contracts@0.2.0-alpha.356
+  - @brains/utils@0.2.0-alpha.356
+  - @brains/templates@0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ### Patch Changes