+++ shared/theme-base/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @brains/theme-base
 
+## 0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ## 0.2.0-alpha.354