+++ shell/core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @brains/core
 
+## 0.2.0-alpha.356
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.356
+  - @brains/image@0.2.0-alpha.356
+  - @brains/operation-context@0.2.0-alpha.356
+  - @brains/site-composition@0.2.0-alpha.356
+  - @brains/utils@0.2.0-alpha.356
+  - @brains/plugins@0.2.0-alpha.356
+  - @brains/recurring-checks@0.2.0-alpha.356
+  - @brains/runtime-state@0.2.0-alpha.356
+  - @brains/scheduler@0.2.0-alpha.356
+  - @brains/templates@0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ### Patch Changes