+++ shell/recurring-checks/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @brains/recurring-checks
 
+## 0.2.0-alpha.356
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/utils@0.2.0-alpha.356
+  - @brains/job-queue@0.2.0-alpha.356
+  - @brains/runtime-state@0.2.0-alpha.356
+  - @brains/scheduler@0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ### Patch Changes