+++ sites/personal/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @brains/site-personal
 
+## 0.2.0-alpha.356
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/blog@0.2.0-alpha.356
+  - @brains/site-info@0.2.0-alpha.356
+  - @brains/profile@0.2.0-alpha.356
+  - @brains/site-composition@0.2.0-alpha.356
+  - @brains/site-engine@0.2.0-alpha.356
+  - @brains/ui-library@0.2.0-alpha.356
+  - @brains/utils@0.2.0-alpha.356
+  - @brains/plugins@0.2.0-alpha.356
+  - @brains/templates@0.2.0-alpha.356
+
 ## 0.2.0-alpha.355
 
 ### Patch Changes